requirements.txt
# Core trading libraries
binance-connector==3.7.0
backtrader==1.9.78.123
TA-Lib==0.4.28

# Data handling
pandas==2.1.4
numpy==1.24.3

# WebSocket and async
websockets==12.0
aiohttp==3.9.1

# Logging and utilities
coloredlogs==15.0.1
python-dotenv==1.0.0
PyYAML==6.0.1

# Plotting for backtest results
matplotlib==3.8.2

# OpenRouter API for AI features
requests==2.31.0